Add 'elein-slime-net-coding-system, which lets elein match up front and back-end coding systems
Given the default "lein swank" :encoding options, and the default value of
'slime-net-coding-system, users are prone to encountering errors such as this one:

https://github.com/technomancy/swank-clojure/issues/57

Documentation is scarce, so in this commit I've tweaked elein such that everything
Just Works out of the box. That is, by simply installing the Marmalade/ELPA packages for
clojure-mode, slime and elein, "lein-swank" will use the correct back-end and net
coding systems without any further customization.